backend/requirements.txt
fastapi
uvicorn
chromadb
numpy
pillow
torch
torchvision
transformers
sentence-transformers
opencv-python
pydantic
python-multipart
scipy
tqdm
pytest
motor